Adds a Python tool for sorting footnote and link references.
These references appear at the bottom of posts, and they tend to
get mangled if I've done a lot of linking and footnoting. I like
to keep them sorted. This is a short Python script that expects
this last part of the document to have been copied into the
pasteboard. It does the sorting, then puts the sorted version back
into the pasteboard.