reformats card lists
This modifies page layout and styling so that the award badges on cards plays better with the styling of .card > img:first-child.

Basically, it allows the corners of the image to be hidden by the rounded card edge (with overflow:hidden) without hiding parts of the badges (with position:absolute).

I did this by wrapping each card in a div.card-container, which has position:relative. I then changed .card to have overflow:hidden.